Safari-611.1.10/LayoutTests/crypto/workers/subtle/rsa-pss-import-key-sign-expected.txt

[Worker] Test signing with RSA-PSS SHA-1 using an imported key in workers

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Starting worker: resources/rsa-pss-import-key-sign.js
PASS [Worker] signature.byteLength is 256
PASS successfullyParsed is true

TEST COMPLETE
